+++ apps/dashboard/src/app/resources/edit/annotation/paragraph-annotation/paragraph-annotation.component.ts
@@ -1,9 +1,27 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  ElementRef,
+  OnDestroy,
+  OnInit,
+  ViewChild,
+} from '@angular/core';
 import { EditResourceService } from '../../edit-resource.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { combineLatest, filter, map, Observable, Subject, switchMap } from 'rxjs';
+import { BehaviorSubject, combineLatest, filter, map, Observable, Subject, switchMap } from 'rxjs';
 import { FieldId, Paragraph, Resource } from '@nuclia/core';
-import { EntityGroup, getParagraphs, ParagraphWithText } from '../../edit-resource.helpers';
+import {
+  EntityAnnotation,
+  EntityGroup,
+  getAnnotatedText,
+  getGeneratedFieldAnnotations,
+  getParagraphAnnotations,
+  getParagraphs,
+  ParagraphWithTextAndAnnotations,
+} from '../../edit-resource.helpers';
+import { takeUntil } from 'rxjs/operators';
+import { SisToastService } from '@nuclia/sistema';
 
 @Component({
   selector: 'app-paragraph-annotation',
@@ -12,6 +30,8 @@ import { EntityGroup, getParagraphs, ParagraphWithText } from '../../edit-resour
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ParagraphAnnotationComponent implements OnInit, OnDestroy {
+  @ViewChild('paragraphsContainer') paragraphsContainer?: ElementRef<HTMLElement>;
+
   unsubscribeAll = new Subject<void>();
 
   private resource: Observable<Resource> = this.editResource.resource.pipe(
@@ -30,16 +50,20 @@ export class ParagraphAnnotationComponent implements OnInit, OnDestroy {
   isModified = false;
   isSaving = false;
 
-  private paragraphsBackup: ParagraphWithText[] = [];
-  paragraphs: ParagraphWithText[] = [];
+  private paragraphsBackup: ParagraphWithTextAndAnnotations[] = [];
+  paragraphs: ParagraphWithTextAndAnnotations[] = [];
 
-  selectedFamily?: EntityGroup;
   entityFamilies: Observable<EntityGroup[]> = this.editResource.loadResourceEntities();
+  selectedFamily: BehaviorSubject<EntityGroup | null> = new BehaviorSubject<EntityGroup | null>(null);
+  selectedEntity?: EntityAnnotation & { paragraphId: string };
+  userSelection = false;
+  buttonPosition?: { top: string; left: string };
 
   constructor(
     private route: ActivatedRoute,
     private router: Router,
     private editResource: EditResourceService,
+    private toaster: SisToastService,
     private cdr: ChangeDetectorRef,
   ) {
     this.route.params
@@ -58,30 +82,41 @@ export class ParagraphAnnotationComponent implements OnInit, OnDestroy {
     this.editResource.setCurrentView('annotation');
 
     // Load paragraphs
-    combineLatest([this.fieldId, this.resource])
+    combineLatest([this.fieldId, this.resource, this.entityFamilies])
       .pipe(
-        map(([fieldId, resource]) => {
+        map(([fieldId, resource, families]) => {
+          const generatedAnnotation: EntityAnnotation[] = getGeneratedFieldAnnotations(resource, fieldId, families);
           const paragraphs: Paragraph[] = getParagraphs(fieldId, resource);
           return paragraphs.map((paragraph) => {
             const paragraphId = this.editResource.getParagraphId(fieldId, paragraph);
-            const enhancedParagraph: ParagraphWithText = {
+            const annotations = getParagraphAnnotations(resource, fieldId, generatedAnnotation, paragraph, families);
+            const paragraphText = resource.getParagraphText(fieldId.field_type, fieldId.field_id, paragraph);
+            const enhancedParagraph: ParagraphWithTextAndAnnotations = {
               ...paragraph,
               paragraphId,
-              text: resource.getParagraphText(fieldId.field_type, fieldId.field_id, paragraph),
+              text: paragraphText,
+              annotatedText: getAnnotatedText(paragraphId, paragraphText, annotations),
+              annotations,
             };
             return enhancedParagraph;
           });
         }),
+        takeUntil(this.unsubscribeAll),
       )
       .subscribe((paragraphs) => {
         this.paragraphsBackup = paragraphs.map((paragraph) => JSON.parse(JSON.stringify(paragraph)));
         this.paragraphs = this.paragraphsBackup.map((paragraph) => JSON.parse(JSON.stringify(paragraph)));
         this.cdr.markForCheck();
+        setTimeout(() => {
+          this.setupMarkListener();
+        });
       });
   }
+
   ngOnDestroy() {
     this.unsubscribeAll.next();
     this.unsubscribeAll.complete();
+    this.cleanUpMarkListener();
   }
 
   save() {
@@ -90,5 +125,60 @@ export class ParagraphAnnotationComponent implements OnInit, OnDestroy {
 
   cancel() {
     // TODO
+  }
+
+  selectFamily(family: EntityGroup) {
+    const selectedFamily = this.selectedFamily.value?.id === family.id ? null : family;
+    this.selectedFamily.next(selectedFamily);
+    this.paragraphs = this.paragraphs.map((paragraph) => ({
+      ...paragraph,
+      annotatedText: getAnnotatedText(paragraph.paragraphId, paragraph.text, paragraph.annotations, selectedFamily),
+    }));
+    setTimeout(() => {
+      this.cleanUpMarkListener();
+      this.setupMarkListener();
+    });
+  }
+
+  private setupMarkListener() {
+    if (this.paragraphsContainer) {
+      this.paragraphsContainer.nativeElement
+        .querySelectorAll('mark[family]')
+        .forEach((mark) => mark.addEventListener('click', this.clickOnAnnotation.bind(this)));
+    }
+  }
+
+  private cleanUpMarkListener() {
+    if (this.paragraphsContainer) {
+      this.paragraphsContainer.nativeElement
+        .querySelectorAll('mark[family]')
+        .forEach((mark) => mark.removeEventListener('click', this.clickOnAnnotation.bind(this)));
+    }
+  }
+
+  private clickOnAnnotation(event: Event) {
+    if (!this.selectedFamily.value) {
+      this.toaster.info('resource.field-annotation.select-family-before');
+      return;
+    }
+    const mark = event.target as HTMLElement;
+    this.selectedEntity = {
+      paragraphId: mark.getAttribute('paragraphId') || '',
+      family: mark.getAttribute('title') || '',
+      token: mark.getAttribute('token') || '',
+      start: parseInt(mark.getAttribute('start') || '0', 10),
+      end: parseInt(mark.getAttribute('end') || '0', 10),
+      klass: mark.getAttribute('family') || '',
+    };
+    const paragraph = mark.parentElement as HTMLElement;
+    const paragraphRect = paragraph.getBoundingClientRect();
+    const markRect = mark.getBoundingClientRect();
+    const buttonBlockWidth =
+      !this.userSelection && this.selectedEntity.klass !== this.selectedFamily.value?.id ? 32 : 16;
+    this.buttonPosition = {
+      left: markRect.height < 16 ? `${markRect.right - paragraphRect.left - buttonBlockWidth}px` : '0',
+      top: `${markRect.bottom - paragraphRect.bottom}px`,
+    };
+    this.cdr.markForCheck();
   }
 }