+++ apps/dashboard/src/app/resources/resources.module.ts
@@ -43,7 +43,7 @@ import {
   SisProgressModule,
   SisStatusComponent,
 } from '@nuclia/sistema';
-import { FileUploadModule } from '@flaps/core';
+import { FileUploadModule, STFPipesModule } from '@flaps/core';
 import { STFSectionNavbarModule } from '../components/section-navbar';
 import { DatasetImportComponent } from './sample-dataset/dataset-import.component';
 import { LoadingModalComponent } from './sample-dataset/loading-modal/loading-modal.component';
@@ -178,6 +178,7 @@ const ROUTES: Routes = [
     PaPopupModule,
     PaChipsModule,
     PaExpanderModule,
+    STFPipesModule,
   ],
   declarations: [...Components],
   exports: [],